requirements.txt
aioredis==1.3.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
Django==4.0.5
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-jazzmin==2.5.0
django-phonenumber-field==7.0.1
django-sms==0.5.0
django-twilio==0.14.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Faker==14.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.4
phonenumbers==8.13.2
Pillow==9.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
twilio==7.15.4
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
urllib3==1.26.12
zope.interface==5.4.0
